+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  + method `add_comment` to class `Site`;
  + method `add_question` to class `Site`;
  + method `edit_answer` to class `Site`;
+ + method `edit_comment` to class `Site`;
  + method `edit_question` to class `Site`;
  + method `get_me` to class `Site`;
  + method `get_my_inbox` to class `Site`;